% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{current_table}
\alias{current_table}
\title{Return the latest employment requirements table, convenience wrapper for table_year defaults}
\usage{
current_table(tablelist = dom_table_list)
}
\value{
The current table for \code{tablelist}.
}
\description{
Return the latest employment requirements table, convenience wrapper for table_year defaults
}
\examples{
current_table()
current_table(tablelist = all_table_list)
}

