% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{imports_table_current}
\alias{imports_table_current}
\title{Return the latest imports employment requirements table, using the convenience wrapper for table_year defaults}
\usage{
imports_table_current()
}
\value{
The current imports table.
}
\description{
Return the latest imports employment requirements table, using the convenience wrapper for table_year defaults
}
\examples{
imports_table_current()
}

