% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{long_impact_df}
\alias{long_impact_df}
\title{Generate a long-format data frame of total, domestic, and import employment in a sector generated by
 million dollars of production of that sector (in 2009 dollars)}
\usage{
long_impact_df(sector_number, prod_output_mm)
}
\arguments{
\item{sector_number}{A \code{sector_number}.}

\item{prod_output_mm}{\code{prod_output_mm}, the amount of final output in $1 million USD.}

\item{year}{The table \code{year}.}
}
\value{
a long-format data frame of total, domestic, and import employment in a sector generated by
  \code{prod_output_mm} dollars production of \code{sector_number} output.
}
\description{
Generate a long-format data frame of total, domestic, and import employment in a sector generated by
 million dollars of production of that sector (in 2009 dollars)
}
\examples{
long_impact_df(27, 10.5)
}

