% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{sector_description}
\alias{sector_description}
\title{Given a sector number, return a description.}
\usage{
sector_description(sector_number, sectors)
}
\arguments{
\item{sector_number}{A Sector Number (1 - 206)}

\item{sectors}{A Vector of Sector Numbers (1 - 206).}
}
\value{
The Industry.Commodity.Description for \code{sector_number} or \code{sectors}.
}
\description{
Given a sector number, return a description.
}
\examples{
sector_description(27)
sector_description(c(27, 28))
}

