% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{sector_direct_impact_all_years}
\alias{sector_direct_impact_all_years}
\title{Returns a vector of the number of jobs (default: domestic only) directly employed in a sector generated by
 $1 million of production of that sector (in 2009 dollars) for all years}
\usage{
sector_direct_impact_all_years(sector, prod_output,
  tablelist = dom_table_list)
}
\arguments{
\item{tablelist}{A \code{tablelist} for the domestic, imports, or total employment tables list.}

\item{sector_number}{A \code{sector_number}.}

\item{prod_output_mm}{\code{prod_output_mm}, the amount of final output in $1 million USD.}
}
\value{
The total number of directly employed for \code{prod_output_mm} dollars production of \code{sector_number} output, \code{year} and \code{tablelist}.
}
\description{
Returns a vector of the number of jobs (default: domestic only) directly employed in a sector generated by
 $1 million of production of that sector (in 2009 dollars) for all years
}
\examples{
sector_direct_impact_all_years(27, 10.5, all_table_list)
}

