% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{sector_emp_impact}
\alias{sector_emp_impact}
\title{Return a vector of the number of jobs by all sectors for a given amount of production of a sector}
\usage{
sector_emp_impact(sector_number, prod_output_mm, year = LATEST_YEAR,
  tablelist = dom_table_list)
}
\arguments{
\item{sector_number}{A \code{sector_number}.}

\item{prod_output_mm}{\code{prod_output_mm}, the amount of final output in $1 million USD.}

\item{year}{The table \code{year}.}

\item{tablelist}{A \code{tablelist} for the domestic, imports, or total employment tables list.}
}
\value{
The number of employed for \code{prod_output_mm} dollars production of \code{sector_number} output, \code{year} and \code{tablelist}.
}
\description{
Return a vector of the number of jobs by all sectors for a given amount of production of a sector
}
\examples{
sector_emp_impact(27, 10.5, 1997, all_table_list)
}

