% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{series_by_sector}
\alias{series_by_sector}
\title{Return the employment requirements coefficient vector for a sector, the number of jobs
generated in all sectors, direct and indirect, by one million dollars output of any
sector-column}
\usage{
series_by_sector(series_table, sector_number)
}
\arguments{
\item{series_table}{A \code{series_table} for the domestic, imports, or total employment tables.}

\item{sector_number}{A \code{sector number}.}
}
\value{
The series table for \code{sector_number}.
}
\description{
Return the employment requirements coefficient vector for a sector, the number of jobs
generated in all sectors, direct and indirect, by one million dollars output of any
sector-column
}
\examples{
series_by_sector(a_series_table, 27)
}

