% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{series_by_sector_by_year}
\alias{series_by_sector_by_year}
\title{Return the employment requirements coefficient vector for a sector and given year, latest year by default}
\usage{
series_by_sector_by_year(sector_number, year = LATEST_YEAR,
  tablelist = dom_table_list)
}
\arguments{
\item{sector_number}{A \code{sector_number}.}

\item{year}{The table \code{year}.}

\item{tablelist}{A \code{tablelist} for the domestic, imports, or total employment tables list.}
}
\value{
The series table for \code{sector_number}, \code{year} and \code{tablelist}.
}
\description{
Return the employment requirements coefficient vector for a sector and given year, latest year by default
}
\examples{
series_by_sector_by_year(27, 1997, all_table_list)
}

