% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_funs.R
\name{table_by_year}
\alias{table_by_year}
\title{Return the employment requirements table for the given year, current by default}
\usage{
table_by_year(year = LATEST_YEAR, tablelist = dom_table_list)
}
\arguments{
\item{year}{A table year, current is default.}

\item{tablelist}{An employment matrix requirements table.}
}
\value{
The table for \code{year} from \code{tabelist}.
}
\description{
Return the employment requirements table for the given year, current by default
}
\examples{
table_by_year()
table_by_year(1997, dom_table_list)
}

